requirements.txt

torch>=2.0.0           
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0  
ctransformers==0.2.5
pinecone-client>=3.0.0
flask
pypdf
huggingface_hub>=0.13.0
sentence-transformers>=2.2.0
langchain>=0.0.225
ipywidgets>=8.0.0
langchain-community
langchain-huggingface
langchain-pinecone

python-dotenv  # to read the .env file folder
-e . # This is the path to the setup.py file of the current project

